Re-reconcile old versions if status has changed
This only applies to automated reconciliations. If a transaction
was previously reconciled and we have decided that it should
be not reconciled, or vice vera, and the older validation was performed
automatically using a previous version of the reconcile engine, we
now update the transaction to the new state.

We now also display the reconcile engine version in the
reconcile dialog, and record the engine version number in the database.