Fix for errors thrown for MQB transactions
This is the first step in supporting MQB transactions. This
commit fixes an error thrown when we receive a user tier
corresponding to a MQB transaction, which is in the format
"XXXX User".

The actual MQB pricing formula is not implemented yet, but
this fix stops the app from crashing when trying to validate
such a transaction.